Convergence: fix RFX model and retune
Output Channel 7 (the rear motor) had the Input applied twice. This
meant that the motor didn't kick in until way late and saturated way
early.

The the front motors had their 200% power multiplier reduced to 70% and
the rear prop was adjusted to be slightly bigger. This brings the
throttle values for hover in line with each-other.

The aircraft had to be retuned after these changes.